+++ src/prob/opf_oltc.jl
@@ -16,6 +16,7 @@ function build_mc_opf_oltc(pm::AbstractUnbalancedPowerModel)
 
     variable_mc_generator_power(pm)
     variable_mc_load_power(pm)
+    variable_mc_storage_power(pm)
 
     constraint_mc_model_voltage(pm)
 
@@ -35,6 +36,13 @@ function build_mc_opf_oltc(pm::AbstractUnbalancedPowerModel)
 
     for i in ids(pm, :bus)
         constraint_mc_power_balance(pm, i)
+    end
+
+    for i in ids(pm, :storage)
+        constraint_storage_state(pm, i)
+        constraint_storage_complementarity_nl(pm, i)
+        constraint_mc_storage_losses(pm, i)
+        constraint_mc_storage_thermal_limit(pm, i)
     end
 
     for i in ids(pm, :branch)
@@ -69,6 +77,7 @@ function build_mc_opf_oltc(pm::AbstractUBFModels)
     variable_mc_transformer_power(pm)
     variable_mc_generator_power(pm)
     variable_mc_load_power(pm)
+    variable_mc_storage_power(pm)
 
     variable_mc_oltc_transformer_tap(pm)
 
@@ -91,6 +100,13 @@ function build_mc_opf_oltc(pm::AbstractUBFModels)
 
     for i in ids(pm, :bus)
         constraint_mc_power_balance(pm, i)
+    end
+
+    for i in ids(pm, :storage)
+        constraint_storage_state(pm, i)
+        constraint_storage_complementarity_nl(pm, i)
+        constraint_mc_storage_losses(pm, i)
+        constraint_mc_storage_thermal_limit(pm, i)
     end
 
     for i in ids(pm, :branch)